Census population by household size per commune in Switzerland between 1970 and 2000

Each row contains the number of households for a municipality, a year and a household size.

Usage

SwissPop

Source

STAT-TAB, section 40.1 "Eidgenössische Volkszählung/1970--2000", item "Privathaushalte nach Haushaltsgrösse und Region".

STAT-TAB URL: http://www.pxweb.bfs.admin.ch/Database/German_40%20-%20Eidgen%C3%B6ssische%20Volksz%C3%A4hlung/40.1%20-%201970-2000/40.1%20-%201970-2000.asp?lang=1&prod=40&secprod=1&openChild=true.

File link: http://www.pxweb.bfs.admin.ch/Database/German_40%20-%20Eidgen%C3%B6ssische%20Volksz%C3%A4hlung/40.1%20-%201970-2000/px-d-40-1A01.px.

Author

Swiss Federal Statistical Office