Create a list of municipality mutations

Municipality mutation lists are required to compute matching tables for municipalities between different points in time.

swc_get_mutations(mids = NULL, canton = NULL)

Arguments

mids
A list of municipality id's (BFS-numbers) of which the mutations should be retrieved.

canton
Canton abbreviation as character (e.g. "GE", "ZH", "TI", etc.) to focus on. If left `NULL` (default) all cantons are considered.

Value

A data frame that represents mutations.

Details

The municipality dataset of the Swiss historicized communes data is a list of snapshots of municipality states. Each state is valid during a certain period of time, adjoining states are linked by admission and abolition numbers: The abolition number of a former state is the same as the admission number of a subsequent state. The states can be thought of as nodes in a graph, where the edges are mutations -- transformations from one state to the next.

This function performs a self-merge on the municipality data: Each abolition number is matched with its corresponding admission number (if available). If no corresponding admission or abolition number is found, the record is included with NAs instead of matched values. Records without admission or abolition number are excluded. The result is a list of mutations, i.e., a list of edges in the graph of municipality state snapshots.

Examples

head(swcGetMutations(), 20)
#> Warning: `swcGetMutations()` was deprecated in munch 0.0.8.
#> Please use `swc_get_mutations()` instead.
#>    mMutationNumber cAbbreviation mHistId.x mId.x mShortName.x mAbolitionMode
#> 1             1000          <NA>        NA    NA         <NA>           <NA>
#> 2             1000          <NA>        NA    NA         <NA>           <NA>
#> 3             1000          <NA>        NA    NA         <NA>           <NA>
#> 4             1000          <NA>        NA    NA         <NA>           <NA>
#> 5             1000          <NA>        NA    NA         <NA>           <NA>
#> 6             1000          <NA>        NA    NA         <NA>           <NA>
#> 7             1000          <NA>        NA    NA         <NA>           <NA>
#> 8             1000          <NA>        NA    NA         <NA>           <NA>
#> 9             1000          <NA>        NA    NA         <NA>           <NA>
#> 10            1000          <NA>        NA    NA         <NA>           <NA>
#> 11            1000          <NA>        NA    NA         <NA>           <NA>
#> 12            1000          <NA>        NA    NA         <NA>           <NA>
#> 13            1000          <NA>        NA    NA         <NA>           <NA>
#> 14            1000          <NA>        NA    NA         <NA>           <NA>
#> 15            1000          <NA>        NA    NA         <NA>           <NA>
#> 16            1000          <NA>        NA    NA         <NA>           <NA>
#> 17            1000          <NA>        NA    NA         <NA>           <NA>
#> 18            1000          <NA>        NA    NA         <NA>           <NA>
#> 19            1000          <NA>        NA    NA         <NA>           <NA>
#> 20            1000          <NA>        NA    NA         <NA>           <NA>
#>    mAbolitionDate mDateOfChange.x mHistId.y mId.y       mShortName.y
#> 1            <NA>            <NA>     11348     1             Aeugst
#> 2            <NA>            <NA>     11742     2 Affoltern am Albis
#> 3            <NA>            <NA>     11801     3         Bonstetten
#> 4            <NA>            <NA>     11992     4    Hausen am Albis
#> 5            <NA>            <NA>     12249     5           Hedingen
#> 6            <NA>            <NA>     12671     6    Kappel am Albis
#> 7            <NA>            <NA>     12674     7             Knonau
#> 8            <NA>            <NA>     12738     8        Maschwanden
#> 9            <NA>            <NA>     12433     9     Mettmenstetten
#> 10           <NA>            <NA>     12497    10           Obfelden
#> 11           <NA>            <NA>     13124    11          Ottenbach
#> 12           <NA>            <NA>     12891    12         Rifferswil
#> 13           <NA>            <NA>     10575    13          Stallikon
#> 14           <NA>            <NA>     11349    14           Wettswil
#> 15           <NA>            <NA>     11735    21            Adlikon
#> 16           <NA>            <NA>     11862    22        Benken (ZH)
#> 17           <NA>            <NA>     11855    23     Berg am Irchel
#> 18           <NA>            <NA>     11480    24     Buch am Irchel
#> 19           <NA>            <NA>     11573    25            Dachsen
#> 20           <NA>            <NA>     12182    26               Dorf
#>             mAdmissionMode mAdmissionDate mDateOfChange.y mMutationDate
#> 1  First-time registration     1960-01-01      1976-11-14    1960-01-01
#> 2  First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 3  First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 4  First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 5  First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 6  First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 7  First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 8  First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 9  First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 10 First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 11 First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 12 First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 13 First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 14 First-time registration     1960-01-01      1976-11-14    1960-01-01
#> 15 First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 16 First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 17 First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 18 First-time registration     1960-01-01      2012-12-31    1960-01-01
#> 19 First-time registration     1960-01-01      1960-01-01    1960-01-01
#> 20 First-time registration     1960-01-01      1960-01-01    1960-01-01
#>        mMutationId
#> 1  1960-01-01.1000
#> 2  1960-01-01.1000
#> 3  1960-01-01.1000
#> 4  1960-01-01.1000
#> 5  1960-01-01.1000
#> 6  1960-01-01.1000
#> 7  1960-01-01.1000
#> 8  1960-01-01.1000
#> 9  1960-01-01.1000
#> 10 1960-01-01.1000
#> 11 1960-01-01.1000
#> 12 1960-01-01.1000
#> 13 1960-01-01.1000
#> 14 1960-01-01.1000
#> 15 1960-01-01.1000
#> 16 1960-01-01.1000
#> 17 1960-01-01.1000
#> 18 1960-01-01.1000
#> 19 1960-01-01.1000
#> 20 1960-01-01.1000
head(subset(swcGetMutations(), !is.na(mHistId.x)), 20)
#> Warning: `swcGetMutations()` was deprecated in munch 0.0.8.
#> Please use `swc_get_mutations()` instead.
#>      mMutationNumber cAbbreviation mHistId.x mId.x          mShortName.x
#> 3103            1001            VD     11248  5512               Bercher
#> 3104            1002            SO     11245  2451            Ichertswil
#> 3105            1002            SO     11244  2455            Lüterkofen
#> 3106            1003            VD     11242  5625      Bussy-sur-Morges
#> 3107            1003            VD     11243  5626 Chardonney-sur-Morges
#> 3108            1004            AG     11240  4021                 Baden
#> 3109            1004            AG     11227  4025               Dättwil
#> 3110            1005            VD     11238  5886    Montreux-Châtelard
#> 3111            1005            VD     11253  5887     Montreux-Planches
#> 3112            1006            GR     11237  3972    Seewis im Prätigau
#> 3113            1007            GR     11236  3881   Conters im Prätigau
#> 3114            1008            FR     11235  2295         Grossbösingen
#> 3115            1009            GR     11234  3982       Disentis/Mustèr
#> 3116            1010            GR     11231  3597                Peiden
#> 3117            1010            GR     11230  3602       Uors (Lumnezia)
#> 3118            1011            SH     11233  2935            Herblingen
#> 3119            1011            SH     11232  2939          Schaffhausen
#> 3120            1012            SG     11229  3404                 Henau
#> 3121            1013            TG     11239  4756     Schönholzerswilen
#> 3122            1013            TG     11266  4757                  Toos
#>                 mAbolitionMode mAbolitionDate mDateOfChange.x mHistId.y mId.y
#> 3103 Reassignment to other d/c     1960-12-31      1960-12-31     13186  5512
#> 3104                 Abolition     1960-12-31      1960-12-31     13187  2455
#> 3105                 Abolition     1960-12-31      1960-12-31     13187  2455
#> 3106                 Abolition     1961-03-31      1961-03-31     13188  5625
#> 3107                 Abolition     1961-03-31      1961-03-31     13188  5625
#> 3108            Change of area     1961-12-31      1961-12-31     13189  4021
#> 3109                 Abolition     1961-12-31      1961-12-31     13189  4021
#> 3110                 Abolition     1961-12-31      1961-12-31     13190  5886
#> 3111                 Abolition     1961-12-31      1961-12-31     13190  5886
#> 3112        Change of name (m)     1961-12-31      1961-12-31     13191  3972
#> 3113        Change of name (m)     1961-12-31      1961-12-31     13192  3881
#> 3114        Change of name (m)     1961-12-31      1961-12-31     13193  2295
#> 3115        Change of name (m)     1962-12-31      1962-12-31     13194  3982
#> 3116                 Abolition     1963-05-23      1963-05-23     13195  3602
#> 3117                 Abolition     1963-05-23      1963-05-23     13195  3602
#> 3118                 Abolition     1963-12-31      1963-12-31     13196  2939
#> 3119            Change of area     1963-12-31      1963-12-31     13196  2939
#> 3120        Change of name (m)     1964-03-09      1964-03-09     13197  3408
#> 3121            Change of area     1964-05-31      1964-05-31     13198  4756
#> 3122                 Abolition     1964-05-31      1964-05-31     13198  4756
#>               mShortName.y            mAdmissionMode mAdmissionDate
#> 3103               Bercher Reassignment to other d/c     1961-01-01
#> 3104 Lüterkofen-Ichertswil                  Creation     1961-01-01
#> 3105 Lüterkofen-Ichertswil                  Creation     1961-01-01
#> 3106      Bussy-Chardonney                  Creation     1961-04-01
#> 3107      Bussy-Chardonney                  Creation     1961-04-01
#> 3108                 Baden            Change of area     1962-01-01
#> 3109                 Baden            Change of area     1962-01-01
#> 3110              Montreux                  Creation     1962-01-01
#> 3111              Montreux                  Creation     1962-01-01
#> 3112   Seewis im Prättigau        Change of name (m)     1962-01-01
#> 3113  Conters im Prättigau        Change of name (m)     1962-01-01
#> 3114              Bösingen        Change of name (m)     1962-01-01
#> 3115       Disentis/Mustér        Change of name (m)     1963-01-01
#> 3116           Uors-Peiden                  Creation     1963-05-24
#> 3117           Uors-Peiden                  Creation     1963-05-24
#> 3118          Schaffhausen            Change of area     1964-01-01
#> 3119          Schaffhausen            Change of area     1964-01-01
#> 3120                 Uzwil        Change of name (m)     1964-03-10
#> 3121     Schönholzerswilen            Change of area     1964-06-01
#> 3122     Schönholzerswilen            Change of area     1964-06-01
#>      mDateOfChange.y mMutationDate     mMutationId
#> 3103      2006-08-31    1961-01-01 1961-01-01.1001
#> 3104      1961-01-01    1961-01-01 1961-01-01.1002
#> 3105      1961-01-01    1961-01-01 1961-01-01.1002
#> 3106      2006-08-31    1961-04-01 1961-04-01.1003
#> 3107      2006-08-31    1961-04-01 1961-04-01.1003
#> 3108      1962-01-01    1962-01-01 1962-01-01.1004
#> 3109      1962-01-01    1962-01-01 1962-01-01.1004
#> 3110      2006-08-31    1962-01-01 1962-01-01.1005
#> 3111      2006-08-31    1962-01-01 1962-01-01.1005
#> 3112      2000-12-31    1962-01-01 1962-01-01.1006
#> 3113      2000-12-31    1962-01-01 1962-01-01.1007
#> 3114      1962-01-01    1962-01-01 1962-01-01.1008
#> 3115      2000-12-31    1963-01-01 1963-01-01.1009
#> 3116      2000-12-31    1963-05-24 1963-05-24.1010
#> 3117      2000-12-31    1963-05-24 1963-05-24.1010
#> 3118      2008-12-31    1964-01-01 1964-01-01.1011
#> 3119      2008-12-31    1964-01-01 1964-01-01.1011
#> 3120      2002-12-31    1964-03-10 1964-03-10.1012
#> 3121      2010-12-31    1964-06-01 1964-06-01.1013
#> 3122      2010-12-31    1964-06-01 1964-06-01.1013